moved spawnPlayer from GameScreen to PortalMapLevel
